+++ ext/libbpf_ruby/libbpf_ruby.c
@@ -13,6 +13,9 @@ static VALUE rb_cLibBPFRubyLink;
 static ID id_ivar_object;
 static ID id_ivar_program;
 static ID id_kwarg_retprobe;
+static ID id_kwarg_pid;
+static ID id_kwarg_offset;
+static ID id_kwarg_func_name;
 
 typedef struct {
   struct bpf_object *bpf_object;
@@ -245,6 +248,34 @@ static VALUE rb_cProgram_attach_kprobe(int argc, VALUE *argv, VALUE self) {
   return libbpf_ruby_link_wrap(self, link);
 }
 
+static VALUE rb_cProgram_attach_uprobe(int argc, VALUE *argv, VALUE self) {
+  VALUE binary_path, kwargs;
+  rb_scan_args(argc, argv, "1:", &binary_path, &kwargs);
+  const char *func_name = NULL;
+  size_t offset = 0;
+  pid_t pid = -1;
+  bool retprobe = false;
+  if (!NIL_P(kwargs)) {
+    VALUE value;
+    value = rb_hash_lookup2(kwargs, ID2SYM(id_kwarg_func_name), Qundef);
+    if (value != Qundef) func_name = StringValueCStr(value);
+    value = rb_hash_lookup2(kwargs, ID2SYM(id_kwarg_offset), Qundef);
+    if (value != Qundef) offset = NUM2SIZET(value);
+    value = rb_hash_lookup2(kwargs, ID2SYM(id_kwarg_pid), Qundef);
+    if (value != Qundef) pid = NUM2INT(value);
+    value = rb_hash_lookup2(kwargs, ID2SYM(id_kwarg_retprobe), Qundef);
+    if (value != Qundef) retprobe = RTEST(value);
+  }
+
+  LIBBPF_OPTS(bpf_uprobe_opts, uopts, .retprobe = retprobe, .func_name = func_name);
+  struct bpf_link *link = bpf_program__attach_uprobe_opts(libbpf_ruby_program_bpf(self), pid, StringValueCStr(binary_path), offset, &uopts);
+  long err = libbpf_get_error(link);
+  if (err) {
+    rb_raise(rb_eRuntimeError, "bpf_program__attach_uprobe_opts failed: %s", strerror(-err));
+  }
+  return libbpf_ruby_link_wrap(self, link);
+}
+
 static VALUE rb_cLink_fd(VALUE self) {
   libbpf_ruby_link_t *libbpf_ruby_link;
   TypedData_Get_Struct(self, libbpf_ruby_link_t, &libbpf_ruby_link_type, libbpf_ruby_link);
@@ -308,6 +339,9 @@ RUBY_FUNC_EXPORTED void Init_libbpf_ruby(void) {
   id_ivar_object = rb_intern("@object");
   id_ivar_program = rb_intern("@program");
   id_kwarg_retprobe = rb_intern("retprobe");
+  id_kwarg_pid = rb_intern("pid");
+  id_kwarg_offset = rb_intern("offset");
+  id_kwarg_func_name = rb_intern("func_name");
 
   VALUE rb_mLibBPFRuby = rb_define_module("LibBPFRuby");
   VALUE rb_cLibBPFRubyObject = rb_define_class_under(rb_mLibBPFRuby, "Object", rb_cObject);
@@ -330,6 +364,7 @@ RUBY_FUNC_EXPORTED void Init_libbpf_ruby(void) {
   rb_define_method(rb_cLibBPFRubyProgram, "attach_tcx", rb_cProgram_attach_tcx, 1);
 #endif
   rb_define_method(rb_cLibBPFRubyProgram, "attach_kprobe", rb_cProgram_attach_kprobe, -1);
+  rb_define_method(rb_cLibBPFRubyProgram, "attach_uprobe", rb_cProgram_attach_uprobe, -1);
 
   rb_undef_alloc_func(rb_cLibBPFRubyLink);
   rb_define_method(rb_cLibBPFRubyLink, "fd", rb_cLink_fd, 0);